+++ lib/src/ui/list_footer/logic.dart
@@ -1,0 +1,12 @@
+part of list_footer;
+
+enum ListFooterState {
+  noMore,
+  loading,
+  failed,
+}
+
+class ListFooterLogic extends GetxController {
+  final state = ListFooterState.noMore.obs;
+  final failedMsg = ''.obs;
+}