+++ lib/src/ui/my_salomon_bottom_bar.dart
@@ -1,0 +1,187 @@
+/// Name: my_salomon_bottom_bar.dart
+///Copyright 2020 Luke Pighetti
+///
+/// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+///
+/// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+///
+/// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+///
+///
+/// Modify by Static4u
+/// Date : 2023/9/13 21:41
+import 'package:base_lib_pub/base_lib_pub.dart';
+import 'package:flutter/material.dart';
+
+class MySalomonBottomBar extends StatelessWidget {
+  /// A bottom bar that faithfully follows the design by Aurélien Salomon
+  ///
+  /// https://dribbble.com/shots/5925052-Google-Bottom-Bar-Navigation-Pattern/
+  const MySalomonBottomBar({
+    Key? key,
+    required this.items,
+    this.backgroundColor,
+    this.currentIndex = 0,
+    this.onTap,
+    this.onDoubleTap,
+    this.selectedItemColor,
+    this.unselectedItemColor,
+    this.selectedColorOpacity,
+    this.itemShape = const StadiumBorder(),
+    this.margin = const EdgeInsets.all(8),
+    this.itemPadding = const EdgeInsets.symmetric(vertical: 10, horizontal: 16),
+    this.duration = const Duration(milliseconds: 500),
+    this.curve = Curves.easeOutQuint,
+  }) : super(key: key);
+
+  /// A list of tabs to display, ie `Home`, `Likes`, etc
+  final List<SalomonBottomBarItem> items;
+
+  /// The tab to display.
+  final int currentIndex;
+
+  /// Returns the index of the tab that was tapped.
+  final Function(int)? onTap;
+
+  /// Returns the index of the tab that was double tapped.
+  final Function(int)? onDoubleTap;
+
+  /// The background color of the bar.
+  final Color? backgroundColor;
+
+  /// The color of the icon and text when the item is selected.
+  final Color? selectedItemColor;
+
+  /// The color of the icon and text when the item is not selected.
+  final Color? unselectedItemColor;
+
+  /// The opacity of color of the touchable background when the item is selected.
+  final double? selectedColorOpacity;
+
+  /// The border shape of each item.
+  final ShapeBorder itemShape;
+
+  /// A convenience field for the margin surrounding the entire widget.
+  final EdgeInsets margin;
+
+  /// The padding of each item.
+  final EdgeInsets itemPadding;
+
+  /// The transition duration
+  final Duration duration;
+
+  /// The transition curve
+  final Curve curve;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+
+    return Container(
+      height: BaseDimens.dBottomBarHeightDefault,
+      color: backgroundColor ?? Colors.transparent,
+      child: SafeArea(
+        minimum: margin,
+        child: Row(
+          /// Using a different alignment when there are 2 items or less
+          /// so it behaves the same as BottomNavigationBar.
+          mainAxisAlignment: items.length <= 2 ? MainAxisAlignment.spaceEvenly : MainAxisAlignment.spaceBetween,
+          children: [
+            for (final item in items)
+              TweenAnimationBuilder<double>(
+                tween: Tween(
+                  end: items.indexOf(item) == currentIndex ? 1.0 : 0.0,
+                ),
+                curve: curve,
+                duration: duration,
+                builder: (context, t, _) {
+                  final _selectedColor = item.selectedColor ?? selectedItemColor ?? theme.primaryColor;
+
+                  final _unselectedColor = item.unselectedColor ?? unselectedItemColor ?? theme.iconTheme.color;
+
+                  return Material(
+                    color: Color.lerp(_selectedColor.withOpacity(0.0), _selectedColor.withOpacity(selectedColorOpacity ?? 0.1), t),
+                    shape: itemShape,
+                    child: InkWell(
+                      onTap: () => onTap?.call(items.indexOf(item)),
+                      onDoubleTap: () => onDoubleTap?.call(items.indexOf(item)),
+                      customBorder: itemShape,
+                      focusColor: _selectedColor.withOpacity(0.1),
+                      highlightColor: _selectedColor.withOpacity(0.1),
+                      splashColor: _selectedColor.withOpacity(0.1),
+                      hoverColor: _selectedColor.withOpacity(0.1),
+                      child: Padding(
+                        padding: itemPadding -
+                            (Directionality.of(context) == TextDirection.ltr
+                                ? EdgeInsets.only(right: itemPadding.right * t)
+                                : EdgeInsets.only(left: itemPadding.left * t)),
+                        child: Row(
+                          children: [
+                            IconTheme(
+                              data: IconThemeData(
+                                color: Color.lerp(_unselectedColor, _selectedColor, t),
+                                size: 24,
+                              ),
+                              child: items.indexOf(item) == currentIndex ? item.activeIcon ?? item.icon : item.icon,
+                            ),
+                            ClipRect(
+                              clipBehavior: Clip.antiAlias,
+                              child: SizedBox(
+                                // height: 24,
+                                child: Align(
+                                  alignment: const Alignment(-0.2, 0.0),
+                                  widthFactor: t,
+                                  child: Padding(
+                                    padding: Directionality.of(context) == TextDirection.ltr
+                                        ? EdgeInsets.only(left: itemPadding.left / 2, right: itemPadding.right)
+                                        : EdgeInsets.only(left: itemPadding.left, right: itemPadding.right / 2),
+                                    child: DefaultTextStyle(
+                                      style: TextStyle(
+                                        color: Color.lerp(_selectedColor.withOpacity(0.0), _selectedColor, t),
+                                        fontWeight: FontWeight.w600,
+                                      ),
+                                      child: item.title,
+                                    ),
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                      ),
+                    ),
+                  );
+                },
+              ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+/// A tab to display in a [SalomonBottomBar]
+class SalomonBottomBarItem {
+  /// An icon to display.
+  final Widget icon;
+
+  /// An icon to display when this tab bar is active.
+  final Widget? activeIcon;
+
+  /// Text to display, ie `Home`
+  final Widget title;
+
+  /// A primary color to use for this tab.
+  final Color? selectedColor;
+
+  /// The color to display when this tab is not selected.
+  final Color? unselectedColor;
+
+  SalomonBottomBarItem({
+    required this.icon,
+    required this.title,
+    this.selectedColor,
+    this.unselectedColor,
+    this.activeIcon,
+  });
+}